Extract DittoSwifttools HeartbeatTool configuration to DittoService so that it keeps running even when you are not on the configuration screen
